Story 25 - board functionality
Functionality in place for
- User sessions:
now forces user to be signed in and stores the username throughout the site Use can logout from the boardview page.

- Board views:
The page to view the lists for each board now accurately shows the lists only for the board selected.

The page to view boards is created listing the boards the user has view or edit permissions on

Login goes directly to board view page now

-Board functionality:
User can share, create new board, and delete board from the board view page. This cascades when you delete the board, it deletes all associated shares